Safari-609.1.11.1/LayoutTests/http/tests/misc/script_onload_delays_window_onload.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<script>
    var scriptLoaded = false;
</script>
<script src="../resources/slow-script.pl?delay=200" onload="scriptLoaded = true;"></script>
<script>
     addEventListener("load", function() {
         shouldBeTrue("scriptLoaded");
     });
</script>
