LayoutTests/http/tests/misc/script_onload_delays_window_onload-expected.txt

PASS scriptLoaded is true
PASS successfullyParsed is true

TEST COMPLETE
